refactor : repository 재설계
