+++ core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/repository/AuthRepository.kt
@@ -1,14 +1,14 @@
 package com.whatever.caramel.core.domain.repository
 
+import com.whatever.caramel.core.domain.vo.auth.AuthResult
 import com.whatever.caramel.core.domain.vo.auth.AuthToken
 import com.whatever.caramel.core.domain.vo.auth.SocialLoginType
-import com.whatever.caramel.core.domain.vo.auth.UserAuth
 
 interface AuthRepository {
     suspend fun loginWithSocialPlatform(
         idToken: String,
         socialLoginType: SocialLoginType,
-    ): UserAuth
+    ): AuthResult
 
     suspend fun refreshAuthToken(oldToken: AuthToken): AuthToken
