+++ core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/repository/CoupleRepository.kt
@@ -1,7 +1,7 @@
 package com.whatever.caramel.core.domain.repository
 
 import com.whatever.caramel.core.domain.entity.Couple
-import com.whatever.caramel.core.domain.vo.couple.Anniversary
+import com.whatever.caramel.core.domain.vo.calendar.Anniversary
 import com.whatever.caramel.core.domain.vo.couple.CoupleInvitationCode
 import com.whatever.caramel.core.domain.vo.couple.CoupleRelationship
 
@@ -27,12 +27,6 @@ interface CoupleRepository {
     ): Couple
 
     suspend fun getCoupleInfo(): Couple
-
-    suspend fun getAnniversaries(
-        coupleId: Long,
-        startDate: String,
-        endDate: String,
-    ): List<Anniversary>
 
     suspend fun deleteCoupleId()
 }
