--- core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/repository/LinkMetadataRepository.kt
@@ -1,7 +1,0 @@
-package com.whatever.caramel.core.domain.repository
-
-import com.whatever.caramel.core.domain.vo.common.LinkMetaData
-
-interface LinkMetadataRepository {
-    suspend fun getLinkMetadata(url: String): LinkMetaData?
-}
