--- core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/repository/TagRepository.kt
@@ -1,7 +1,0 @@
-package com.whatever.caramel.core.domain.repository
-
-import com.whatever.caramel.core.domain.entity.Tag
-
-interface TagRepository {
-    suspend fun getTags(): List<Tag>
-}
